refactor logic out so we can re-use it in the arm(s)
